Parcels: Generate cdh_alternatives.sh script
The alternatives configuration script is responsible for setting
up system-wide symlinks for program binaries/scripts and config
directories that point into the parcel directory.

This allows the user to invoke the usual client applications without
having to install rpm/deb system packages.

Cloudera Manager is responsible for invoking the script when a given
parcel is activated or deactivated.

For consistency, all binaries and config directories are configured,
even those which should be completely irrelevant due to CM (such as
zookeeper conf dir and the zookeeper-server script). The priority
is kept low because any user configured alternatives for these
should always take precedence (and we don't want to override any
CM generated alternatives either)

Finally, the nature of alternatives is to do nothing if the target
location is already something not managed by alternatives. This is
desirable as we want to defer to any installed system packages.
(cherry picked from commit cb9a355fd564922fe5a7248975c601c5beca8f9d)