Parcels: Turn on active parcel symlink for Cloudera components.
In response to initial customer feedback on parcels, it seems pretty
clear that we need to make it easier for them to use paths within
the parcel, as much as we'd like to avoid it.

To help with that, this change turns on the active parcel symlink.

This is a parcel feature that we added for Informatica, but which
now can be seen to be useful for CDH as well. It allows a user to
construct a path to the active parcel that remains stable when
the active parcel changes (eg: /opt/cloudera/parcels/CDH/). This
can be used by external code wanting to add CDH locations to the
classpath, or in documentation that refers to jars (eg: example
and streaming hadoop jars)

After consideration, we are rejecting adding /usr/lib symlinks as
these open up a huge can of worms with respect to dealing with
users placing files in these locations and how to handle packages
being installed at the same time.
(cherry picked from commit eee7847d0bb434eb437a18739ca0acb36d05e028)