[Parcels] Add parcel 'provides' tags for CDH and Solr.
In CM5 (or maybe 4.7) we are introducing an extensible mechanism
for service handlers to identify parcels that are relevant to their
interests. This cannot be done through service handlers enumeration
parcel names, as third party parcels can and will exist that provide
plugins and extensions for services, so we need a separate way for
a parcel to publish tags that describe what it provides.

I do not want these to be tightly coupled to component names, so
we're introducing a separate array for them.

The initial intention is that every service will look for 'foo' and
'foo-plugin' tags, with 'cdh' and 'cdh-plugin' being special and
expanded inside CM.

In this change we see that the CDH parcel provides 'cdh',
unsurprisingly, while solr provides 'solr', and more notably,
plugins for flume and hbase. Subsequent changes will cover Impala
and gplextras.