CLOUDERA-BUILD. Moving to building true source packages without binaries.
There's a lot of work in this, but the gist is the switch from source
packages that are derived from the binary tarball to source packages
deriving from a clean source tarball. So do-component-builds now
deliberately flip between either the normal behavior so far (building
a binary tarball, basically) or just building a source tarball, if the
GIT_REPO environment variable is set - that's set when run through the
Makefile in cdh.git, but won't be set when we run on a package build
slave, etc.

This also resulted in various changes to paths and the like - a number
of packages were pulling bits from odd places, but now everything's consistent.