[parcels] OPSAPS-16967: Add alternatives.json for Spark parcel
The Spark parcel for CDH4 is intended to be used with CM4 and CM5.
While we could just provide the legacy alternatives script and rely
on whitelisting in CM5, it seems better to get it right first time
and ship both forms so that we don't need spark in the whitelist.