CDH-16922. [spark] Set reasonable worker dir and log dir
When using start-all.sh, Spark creates the worker dir and log dir if they don't
already exist. The default worker dir is no good. So we have to specify it.

I'm picking /var/run/spark/work, which makes it transient. Note that
Spark stores the per-application stderr/stdout (per executor) in the worker
dir. So one may argue that it's not transient runtime data. But these are very
noisy logs, and if anybody cares, they should move/aggregate them to somewhere
else, instead of leaving them scattered on every node.