[spark] CDH-22331. Do not add examples jar to classpath
The examples jar overrides things it should, and ends up breaking
things. Since it's not needed, don't add it to every application's
classpath.

This also allows running the examples succesfully, since spark-submit
will add it to the classpath after all the other Hadoop/Spark paths,
so it won't be overriding anything.