[spark] CDH-23731. Hardcode Hadoop classpath.
Invoking "hadoop classpath" doesn't work when you have packages and
parcels installed on the same system; the package location is always
picked up for some reason. To avoid that, build the classpath manually
using the env variables, and append the result of "hadoop classpath"
just in case it adds something we're missing.