Kudu: changes to accommodate client library versioning
With SONAME version symlinks now part of the Kudu client library, we've
created the following files (for each build type):

  lrwxrwxrwx 1 adar adar       19 Aug 13 00:00 libkudu_client.so -> libkudu_client.so.0
  lrwxrwxrwx 1 adar adar       23 Aug 13 00:00 libkudu_client.so.0 -> libkudu_client.so.0.1.0
  -rwxrwxr-x 1 adar adar 40257832 Aug 13 00:00 libkudu_client.so.0.1.0

According to the Debian packaging guide [1], the library's runtime package
should include the library and its SONAME symlink (i.e. libkudu_client.so.0)
while the development package should include the non-versioned symlink
(i.e. libkudu_client.so). For lack of better guidance, I've applied this
convention to the RPMs as well.

The packaging guide also mentions that the SONAME should be part of the
package name, so that multiple packages with different SONAMEs can be
installed simultaneously. This seems like a good thing to do for RPMs.

Lastly, the debug/release distinction is even more annoying with SONAME
symlinks. I settled on putting the release 'file set' in the toplevel lib
directory, and the debug set in a subdirectory.

1. https://www.debian.org/doc/debian-policy/ch-sharedlibs.html