KUDU-945, KUDU-990: lots and lots of packaging cosmetic changes
This patch applies a bunch of different cosmetic changes to the Kudu
packages. Details:
- Add per-server gflagfiles to the kudu-conf alternatives directory. It's a
  little messy since the init scripts read some of the variables but I
  didn't feel like parsing the gflagfiles there. The end result is a dance
  between the gflagfile and the files in /etc/default. These new files
  belong to the kudu-master and kudu-tserver packages.
- Print some useful text from the init scripts if stuff isn't found.
- Use ts-cli to return from the init script only when after sending an
  RPC to the server successfully.

After some repeated changes to the init script templates, I got tired of
generating them via the Bigtop templating mechanism and replaced them with
verbatim scripts. We still do a little templating (for chkconfig), but
it's a simple procedure with sed. I think the result is clearer overall.