CDH-34616. Avoiding a dependency loop between Search and Sentry.
Sentry requires Solr and Solr requries Sentry, which
makes it hard to get the build going. More specifically,
contrib/sentry-handlers exists to pull in some
sentry-bindings from sentry into the search distribution.
These are not necessary in the "make solr-maven" phase
(because it's not an interesting Maven artifact), and
skipping them allows for some incremental goodness.

This change does nothing until a related solr ant change
goes in, but it's harmless and not blocked on that.