CDH-34616. Pushing $M2_REPO_SUFFIX through packaging scripts.
A recent change to cdh.git lets me build things in parallel
on a multi-core machine.  However, Maven doesn't like two
processes working on ~/.m2/repository at the same time.  To
work around this, I've configured ~/.m2/settings.xml
to apply a suffix to the local repository, which is set
by the build (makefile) that uses this stuff.  This change also
passes those changes through the build scripts that
specify the repository manually.

Change-Id: I727ace41455f4c777619ad712d389ea8a5786ef2