RELENG-125. Concessions in generic packaging code for "cdh6.x" versions.
To avoid needing to rename the non-release "cdh6.x" (or "cdh6.3.x")
branches, we wish to name them "cdh6.x".  This confuses the generic
version-less symlink code, because it's looking to strip things
like "cdh1.2.3" but refuses to strip away "cdh6.x".  Long-story
short, this produced excruciatingly opaque error messages like
so, where the duplicate "cdh6.x" is clearly suspect.

To address this, we manipulate the regular expressions
to work around this.  Fortunately, there were unit
tests here and I was able to run them with "make".

    + ln -s ../../lib/spark-assembly-1.6.0-cdh6.x-SNAPSHOT-hadoop2.6.0-cdh6.x-SNAPSHOT.jar ../../lib/spark-assembly-cdh6.x-cdh6.x.jar ./spark-assembly.jar
    ln: target `./spark-assembly.jar' is not a directory