Fix for Impala do-component-build overcleaning the workspace.
Impala's "clean_cmake" code removes all cmake-looking files
from $WORKSPACE.  Though this works if your workspace only
contains Impala files for your current build, it is wrong
if your workspace also contains other components which
use CMake. I experienced this directly in a build which
was concurrently (in the guise of the same Jenkins job)
building both Hadoop and Impala, and Impala's do-component-build
was cleaning Hadoop's files.

$WORKSPACE is set by Jenkins and is not set in other contexts.
Find defaults to searching in the current directory if no
paths are specified.  This has masked this subtlety, because
if $WORKSPACE isn't set, it's doing exactly what it should,
namely cleaning up underneath the current working directory.
(This script already assumes it's being run from the root
of the Impala tree; bin/bootstrap_toolchain.sh is being
run, for example.)

Interestingly, even under Jenkins, the Debian build would
have cleared $WORKSPACE, because Debian packaging builds
have an explicit whitelist of variables that pass through.

In this change, I replaced $WORKSPACE with "." to make
it obvious that we're working with the current directory.

Change-Id: I0da74789cb54bf1f4e9859b6b0bfae351bd79762