[CDH-42306] Configure Solr watchdog to always use FQDN instead of 127.0.0.1
For Debian_8 + Java_8 combination in secure enviornment (kerberos + TLS), Solr
server startup was failing due to watchdog error (HTTP 600). Ideally we should
be using FQDN in the kerberized environment. Hence updated the packaging to use
the value of SOLR_HOSTNAME environment variable (which is used by Solr to advertise
server hostname for the clients). If this variable is unset, the using FQDN as the
default value (instead of 127.0.0.1).