CDH-57040. flume-ng-agent script never perform hard kill of agent after stop / restart
This change fixes a bug in flume-ng-agent init script.

There was a kill -0 command, which tested if the process is still running.
It was bracketed to bash test command, which was by mistake.

This test is not necessary at all, kill -KILL can be emitted even if the
process is not alive anymore, since the output is redirected to /dev/null.