CDH-58997: Enable filtering out libjvm.so dependency of hadoop-lzo rpm
By overriding _use_internal_dependency_generator to 0, the previously
added filtering code will be effective. As a side effect, in RedHat 6
rtld(GNU_HASH) is also removed, so it is explicitely re-added.

Removing libjvm.so will remove an implicite OpenJDK dependency install,
which is an unsupported Java platform as of now.

Tested on RedHat 7.4 and 6.9, with default rpm and rpmbuild packages,
libjvm.so dependency is now eliminated.

NOTE: From rpm version 4.9 a new method is implemented to add explicit
excludes to the requirements list. However in RedHat 6 the default
version of rpm is 4.8. See.:
http://fedoraproject.org/wiki/Packaging:AutoProvidesAndRequiresFiltering