CDH-79562: Package the release build of Impala-lzo
We have always packaged the debug build of Impala-lzo. This was not
a problem until the debug build of Impala-lzo became dependent on
the gutil pibrary as an unintended consequence of fixing IMPALA-7272
for Impala. Unfortunately the new dependency it introduced leaked
over to Impala-lzo.

The patch changes the packaging code to build Impala and Impala-lzo
in release mode to package the release binary, which eliminates
the gutil dependency.